+++ src/App.js
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import React from "react";
-import Nav from "./common/Nav";
+import Nav from "./components/common/Nav";
 import styled from "styled-components";
 
 const Contents = styled.div`
