+++ src/components/Landing/LandingNav.jsx
@@ -1,12 +1,12 @@
 import Logo_pandaMarketText from "../../assets/root_img/Logo_PandaMarketText.svg";
-import useWindowSize from "../../Hooks/useWindowSize";
-import Button from "../../common/Button";
+// import useWindowSize from "../../hooks/useWindowSize";
+import Button from "../../components/common/Button";
 import { Link } from "react-router-dom";
 import { Contents, Logo } from "./NavStyle.jsx";
 
 //
 function LandingNav() {
-  const device = useWindowSize();
+  // const device = useWindowSize();
   return (
     <Contents>
       <Link to="./">
